Feature/heroku (#1)
* Adds support for Heroku

* Adds some React generated files to gitignore

* Added package.json to ensure that Heroku can run npm

* Bump the node and npm versions for Heroku

* Adds loading of application key from environment variable

* adds api prefix to config

* Fixes wrong use of short-hand for application.secret

* Fixes rout to handler for root

* Ensures .gitignore includes .env* files

* Adds host filter

* Fixed typo in play.filter of application.conf

* Complete overhaul client interface

Formerly the client was generated with NPM and React. This has now be
refactored to Scala.js with React. The client and server have been
pulled into separate subprojects with a shared component.

* Complete overhaul client interface

Formerly the client was generated with NPM and React. This has now be
refactored to Scala.js with React. The client and server have been
pulled into separate subprojects with a shared component.

* Adds application secret retrieval from ENV

* Resolves an issue that caused heroku not to find the executable